MRC-143 Separated FullResultsPageHelper
