+++ app/views/FullResultsPageView.scala.html
@@ -18,6 +18,7 @@
 @import forms._
 @import utils.CurrencyUtils
 @import views.helpers.ResultsPageHelper._
+@import views.helpers.FullResultsPageHelper._
 @import forms.DateUtils._
 @import java.time.LocalDate
